removing schemas to separated folder
