+++ src/components/contactUs/ContactUs.tsx
@@ -2,19 +2,11 @@ import { useContext, useRef } from 'react'
 import Input from '../shared/Input'
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { ThemeContext } from '../features/ThemeFeature/ThemeProvider';
-import Joi from "joi";
-import { tlds } from '@hapi/tlds';
 import { joiResolver } from '@hookform/resolvers/joi';
 import { ContactUsMessage } from '../../types/types';
 import useAxios from '../customHooks/useAxios';
 import { toast } from 'react-toastify';
-
-const schema = Joi.object({
-  fullName: Joi.string().min(4).max(32),
-  email: Joi.string().email({ tlds: { allow: tlds } }).min(7).max(64),
-  subject: Joi.string().min(4).max(32),
-  message: Joi.string().min(4).max(256),
-})
+import { contactUsSchema } from '../../schemas/contactUsSchema';
 
 function ContactUs() {
   const loadingToastRef = useRef(null)
@@ -26,7 +18,7 @@ function ContactUs() {
     handleSubmit,
     formState: { errors, isSubmitting },
   } = useForm<ContactUsMessage>({
-    resolver: joiResolver(schema)
+    resolver: joiResolver(contactUsSchema)
   });
