+++ src/components/singleProduct/SingleProductPageComponents/EditReview.tsx
@@ -3,7 +3,6 @@ import { ThemeContext } from '../../features/ThemeFeature/ThemeProvider';
 import Rating from '@mui/material/Rating';
 import "./singleProductTabs.css"
 import { SubmitHandler, useForm } from 'react-hook-form';
-import * as yup from "yup"
 import { yupResolver } from '@hookform/resolvers/yup';
 import { addReview } from '../../../types/types';
 import useAxios from '../../customHooks/useAxios';
@@ -15,11 +14,7 @@ import { toast } from 'react-toastify';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { ThemeProvider } from '@mui/material/styles';
 import { createTheme } from '@mui/material/styles';
-
-const schema = yup.object({
-    comment: yup.string().min(4).max(256).required(),
-    rating: yup.number().oneOf([0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5]).required()
-})
+import { editReviewSchema } from '../../../schemas/reviewSchemas';
 
 function EditReview({ setIsEditReviewModalOpen,review }) {
     const { PUT } = useAxios()
@@ -52,7 +47,7 @@ function EditReview({ setIsEditReviewModalOpen,review }) {
         trigger,
         reset,
     } = useForm<addReview>({
-        resolver: yupResolver(schema),
+        resolver: yupResolver(editReviewSchema),
     })
 
     const onSubmit: SubmitHandler<addReview> = async (submittedData) => {
